% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetacommunityDynamicsFctsOikos.R
\name{bevHoltC}
\alias{bevHoltC}
\title{Function definition for Beverton-Holt competition model}
\usage{
bevHoltC(N, a)
}
\description{
Function definition for Beverton-Holt competition model
}
