% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{bevHoltFct}
\alias{bevHoltFct}
\title{Beverton-Holt function}
\usage{
bevHoltFct(R, N, A, alpha)
}
\arguments{
\item{R}{a vector of species growth rates}

\item{N}{a vector of species population sizes}

\item{A}{a scalar for competition coefficient}

\item{alpha}{the competition coefficient see \code{\link[=alphaterm]{alphaterm()}} for its
computation}
}
\description{
To simulate growth easily, use the Beverton-Holt equation (ref?)
}
