% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{check_trait_weights}
\alias{check_trait_weights}
\title{Check trait weights data.frame}
\usage{
check_trait_weights(trait_weights, traits)
}
\arguments{
\item{trait_weights}{data.frame with at least three columns equal to \code{trait}
(giving the name of the concerned traits in \code{traits}
df), \code{growth_weight} the relative weight of the trait in
growth and \code{compet_weight} the relative weight of the
trait in competition.}

\item{traits}{a species-traits data.frame with species as rownames and
traits as numeric columns with names matching
\code{trait_weights} column \code{trait}}
}
\value{
nothing if data.frame passess the checks, stops early otherwise.
}
\description{
This is an internal help function to check the trait weights data.frame
(used in \code{\link[=multigen]{multigen()}}). The structure of the given trait weights is fixed:
one column should be named \code{trait`` with the names of the traits in it. The two other columns }growth_weight\code{and}compet_weight\code{contains respectively the relative weight of the trait in growth and competition. The function also additionnally check that the traits in the}trait\code{column are in the}traits` data.frame.
}
\examples{
# Working trait weights data.frame
traits = data.frame(trait1 = 1, trait2 = 2, trait3 = 3)

weight_1 = data.frame(
   trait = c("trait1", "trait2", "trait3"),
   growth_weight = c(0.5, 0.5, 0),
   compet_weight = c(0, 0.5, 0.5))

# Silent function
check_trait_weights(weight_1, traits)
\dontrun{
# Not valid trait weights data.frame
not_valid = data.frame(trait = c("trait1", "trait2", "trait3"),
    growth_weight = c(0.5, 0.8, 0),
    compet_weight = c(0, 0.5, 0.9))

# Stop and error
check_trait_weights(not_valid, traits)
}
}
