% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{compute_hierarchical_compet}
\alias{compute_hierarchical_compet}
\title{Compute Hierarchical Competition coefficient at each time step}
\usage{
compute_hierarchical_compet(composition_given_time_step, trait_values,
  trait_weights, H, th_max, th_min)
}
\arguments{
\item{composition_given_time_step}{composition matrix at a given time step
(a site-species matrix with sites in rows)}

\item{trait_values}{a trait matrix}

\item{trait_weights}{a scenario data.frame}

\item{H}{the hierarchical competition scalar}

\item{th_max}{the maximum hierarchical trait value (in general max. trait in
the pool)}

\item{th_min}{the minimum hierarchical trait value (in general min. trait in
the pool)}
}
\description{
Outputs a matrix of additional growth per patch per species given by
hierarchical competition. The values are considered
}
