% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weighted_performance.R
\name{compute_weighted_performance}
\alias{compute_weighted_performance}
\title{Community Performance Estimators}
\usage{
compute_weighted_performance(perf_df, trait_df)
}
\arguments{
\item{perf_df}{species performance data frame output by
\code{\link[=extract_performances_from_simul]{extract_performances_from_simul()}}}

\item{trait_df}{trait data.frame containing columns \code{trait_cor}, \code{seed} and
\code{species}}
}
\description{
Compute for each community all traits:
\itemize{
\item weighted mean of trait by species purely environmental growth rate
\code{pur_env},
\item weighted mean of trait by species realized growth rate \code{weighted_growth},
\item weighted mean of trait by species abundance \code{cwm},
\item trait of species with highest growth rate per community \code{best_growth},
\item trait of species with highest abundance per community \code{best_abund},
\item other community weighted moments such as variance, skewness and kurtosis.
}
}
