% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_traits.R
\name{create_trait_weights}
\alias{create_trait_weights}
\title{Generates a data.frame of trait weights}
\usage{
create_trait_weights(R, A, H)
}
\arguments{
\item{R}{an integer value giving the contribution of trait in growth}

\item{A}{an integer value giving the contribution of trait in limiting
similarity}

\item{H}{an integer value giving the contribution of trait in hierarchical
competition}
}
\description{
Only in the special case of 3 traits with 1 always driving growth and
another one only driving competition
}
\examples{
create_trait_weights(50, 50, 0)
}
