% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{env_curve}
\alias{env_curve}
\title{Species growth rate for a given trait and environment}
\usage{
env_curve(trait_values, env_value, trait_weights, k = 2, width = 0.5,
  H = 0, th_max = 25, th_min = 1, h_fun = "sum")
}
\arguments{
\item{trait_values}{a numeric vector of species trait values}

\item{env_value}{a single numeric value giving the environmental variable}

\item{trait_weights}{data.frame with at least three columns equal to \code{trait}
(giving the name of the concerned traits in \code{traits}
df), \code{growth_weight} the relative weight of the trait in
growth and \code{compet_weight} the relative weight of the
trait in competition, both hierarchical and based on
limiting similarity.}

\item{k}{a scalar giving the maximum growth rate in optimal environment}

\item{width}{a numeric for niche breadth, constant in gaussian function}

\item{H}{a numeric for hierarchical competition such as H/k <= 1}

\item{th_max}{a numeric for the hierarchical trait value maximizing
hierarchical competiton}

\item{th_min}{a numeric for the hierarchical trait value minimizing
hierarchical competition}

\item{h_fun}{a character name of a function that describes how hierarchical
is combined to environmental-based growth (default: \code{"sum"})}
}
\description{
Using traits that affect growth rate and specified environments, this
function returns a numeric value of expected growth rates given the traits
and the environmental value. The total growth rate is then the average of the
growth rates computed with each trait.
}
\details{
For the moment the environmental filter follows a Gaussian distribution:

\deqn{
    R_{i, x} = k \times \exp(- \frac{(trait_i - env_x)^2}{2\times {width}^2})
}{
    R_ix = k * exp((t_i - env_x)^2/(2 * width^2))
},
where t_i is trait of species i, env_x the environmental value in patch x, k
a scalar giving the maximal growth rate and width the environmental breadth
of species.
}
