% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{envtrtcurve}
\alias{envtrtcurve}
\title{Species growth rate for a given trait and environment}
\usage{
envtrtcurve(trti, envx, types, k = 2, c = 0.5)
}
\arguments{
\item{trti}{a data.frame of species' traits}

\item{envx}{a data.frame}

\item{types}{a vector of trait types indicating if traits should be used to
compute growth rate (types \code{RA} or \code{R})}

\item{k}{scalar for growth rate}

\item{c}{constant in gaussian function (standard deviation)}
}
\description{
Using traits that affect growth rate and specified environments, this
functionr returns a data.frame of expected growth rates given the traits and
the environmental value. Suppose a gaussian relationship between growth rate
and environmental value. The total growth rate is then the average of the
growth rates computed with each trait.
}
