% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_growth_rates.R
\name{extract_growth_rates}
\alias{extract_growth_rates}
\title{Extract different growth rates from fdcoexist simulation}
\usage{
extract_growth_rates(simul, chosen_time = NULL)
}
\arguments{
\item{simul}{Simulation object from \code{\link[=multigen]{multigen()}}}

\item{chosen_time}{numeric timestep at which to extract growth rates}
}
\description{
Extract different growth rates from fdcoexist simulation
}
