% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mismatches.R
\name{extract_mismatches}
\alias{extract_mismatches}
\title{Extract species mismatches}
\usage{
extract_mismatches(x, z)
}
\arguments{
\item{x}{Simulation object from \code{\link[=multigen]{multigen()}}}

\item{z}{Number of the species}
}
\description{
Extract species mismatches
}
