% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_performances_from_simul.R
\name{extract_performances_from_simul}
\alias{extract_performances_from_simul}
\title{Extract Performance Estimators From Simulations}
\usage{
extract_performances_from_simul(simul, trait_list,
  realized_growth_rate = TRUE)
}
\arguments{
\item{simul}{[\code{list(1+)}]\cr{}
single simulation from \code{\link[=multigen]{multigen()}}}

\item{trait_list}{[\code{list(1+)}]\cr{}
a nested list of trait ordered by seeds used as well as
trait correlation name}

\item{realized_growth_rate}{[\code{boolean(1)}]\cr{}
should realized growth rate be computed}
}
\description{
Extract Performance Estimators From Simulations
}
