% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{extract_trait_contrib_from_scenar}
\alias{extract_trait_contrib_from_scenar}
\title{Extract Contribution from Scenario Name}
\usage{
extract_trait_contrib_from_scenar(scenario)
}
\arguments{
\item{scenario}{a character of type \code{"R0A0H0"}}
}
\value{
a list with three elements:
\itemize{
\item \code{R} an integer giving the contribution of trait 2 to growth
\item \code{A} an integer giving the contribution of trait 2 to competition (limiting
similarity)
\item \code{H} an integer giving the contribution of trait 2 to hierarchical
competition
}
}
\description{
All scenarios names are formatted the same way 'RXAYHZ' with X, Y and Z being
integers between 0 and 100
}
