% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cor_traits_rand.R
\name{generate_cor_traits_rand}
\alias{generate_cor_traits_rand}
\title{Generate random traits
Compared to generate_cor_traits() introduce a little of variability in first
trait as instead of being directly determined by the species number it adds
little white noise to it and scale it to a minimum of 0 if negative and
maximum of 25 if maximum value is over 25.}
\usage{
generate_cor_traits_rand(
  number_patches,
  number_species,
  number_other = 9,
  cor_coef = 0.7,
  min_value = 1
)
}
\description{
Generate random traits
Compared to generate_cor_traits() introduce a little of variability in first
trait as instead of being directly determined by the species number it adds
little white noise to it and scale it to a minimum of 0 if negative and
maximum of 25 if maximum value is over 25.
}
