% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_traits.R
\name{generate_traits}
\alias{generate_traits}
\title{Generate Traits data.frame}
\usage{
generate_traits(n_species, min_val, max_val)
}
\arguments{
\item{n_species}{the number of species to generate}

\item{min_val}{minimum trait value (for all traits)}

\item{max_val}{maximum trait value (for all traits)}
}
\description{
Generates a random traits data.frame by default the traits are uniformly
& independently distributed with specifiex minimum and maximum values
}
