% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{landscape_df}
\alias{landscape_df}
\title{Simpler data.frame extracted from population}
\usage{
landscape_df(given_composition, time = 150)
}
\arguments{
\item{given_composition}{a result of a call to \code{multigen()}}

\item{time}{the timestep at which population abundances should
be extracted}
}
\value{
a data.frame with the following columns
* \code{patch} the label of the patch
* \code{species} the name of the species
* \code{N} the local abundance of species in given patch
* \code{patch_optim} environmental value of given patch
}
\description{
Simpler data.frame extracted from population
}
