% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_simul.R
\name{meta_simul}
\alias{meta_simul}
\title{Simulations with various parameters}
\usage{
meta_simul(seed_number, given_k = k, given_A = A, given_B = B,
  given_H = H, given_scenars = NULL, given_traits = NULL,
  given_h_fun = "sum", given_di_thresh = 24, given_env,
  given_composition, given_d, given_env_width = 5)
}
\arguments{
\item{seed_number}{the actual seed used to generate traits}

\item{given_k}{the maximum achievable growth}

\item{given_A}{the intensity of limiting similarity}

\item{given_B}{the intensity of intra-specific competition}

\item{given_H}{the intensity of hierarchical competition}

\item{given_scenars}{if provided the data.frame of scenarios}

\item{given_traits}{if provided the traits data.frame to use otherwise
generate uncorrelated traits}

\item{given_h_fun}{function to use to combine hierarchical competition}

\item{given_di_thresh}{the threshold distance at which species are
considered maximally dissimilar and do not compete anymore in
limiting similarity}

\item{given_env}{the environmental width of the different patches}

\item{given_composition}{the empty matrix with good dimensions to be used
to follow population dynamics}

\item{given_d}{the index of dispersal}
}
\description{
Simulations with various parameters
}
