% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismatch.R
\name{mismatch}
\alias{mismatch}
\title{Plot mismatch per species between environmental optimum and max. abundance}
\usage{
mismatch(simul, n_patches = 25, sp, time = 50, plot = TRUE)
}
\arguments{
\item{simul}{results array of a given simulation (a \code{\link[=multigen]{multigen()}} output)}

\item{n_patches}{an integer indicating the patch number}

\item{sp}{integer for the number of species to consider}

\item{time}{an integer indicating the number of time steps}

\item{plot}{a boolean determining whether to plot or not the mismatch}
}
\description{
Plot mismatch per species between environmental optimum and max. abundance
}
