% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{multigen}
\alias{multigen}
\title{Function to run the simulation}
\usage{
multigen(traits, trait_weights, env, time, species, patches, composition,
  A, d, k, width)
}
\arguments{
\item{traits}{a species-traits data.frame with species as rownames and
traits as numeric columns with names matching
\code{trait_weights} column \code{trait}}

\item{trait_weights}{data.frame with at least three columns equal to \code{trait}
(giving the name of the concerned traits in \code{traits}
df), \code{growth_weight} the relative weight of the trait in
growth and \code{compet_weight} the relative weight of the
trait in competition.}

\item{env}{a vector of environmental values}

\item{time}{an integer giving the total number of generations}

\item{species}{an integer giving the total number of species to simulate}

\item{patches}{an integer giving the total number of patches to simulate}

\item{composition}{the actual array containing species abundances per site
over time, giving the initial populations of each species}

\item{A}{the scalar of competition coefficent (see \code{\link[=bevHoltFct]{bevHoltFct()}})}

\item{d}{a numeric value between 0 and 1 giving the percentage of
dispersal occurring across all patches}

\item{k}{a scalar giving the maximum growth rate in optimal
environment}

\item{width}{a numeric giving niche breadth of all species}
}
\description{
Using specified parameters this function run the simulation
}
