% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patch.R
\name{plot_patch}
\alias{plot_patch}
\title{Plot the dynamics of a single patch}
\usage{
plot_patch(results, patch, time)
}
\arguments{
\item{results}{results array of a given simulation}

\item{patch}{patch number}

\item{time}{time number}
}
\description{
Plot the dynamics of a single patch
}
