% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rh.R
\name{plot_rh}
\alias{plot_rh}
\title{Plot hierarchical growth of species in all the patches}
\usage{
plot_rh(simul, n_patches, time)
}
\arguments{
\item{simul}{results array of a given simulation (a \code{\link[=multigen]{multigen()}} output)}

\item{n_patches}{an integer indicating the patch number}

\item{time}{an integer indicating the number of time steps}
}
\description{
Automatically assigns one color per species.
}
