% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_env_growth.R
\name{r_env}
\alias{r_env}
\title{Plot environmental response curves of species in all the patches}
\usage{
r_env(simul, n_patches, sp, plot = TRUE)
}
\arguments{
\item{simul}{results array of a given simulation (a \code{\link[=multigen]{multigen()}} output)}

\item{n_patches}{an integer indicating the patch number}

\item{sp}{integer for the number of species to consider}

\item{plot}{a boolean determining whether to plot or not the response curves}
}
\description{
Automatically assigns one color per species.
}
