% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_env_CT.R
\name{r_env_CT}
\alias{r_env_CT}
\title{Extract Species by Patch Growth Rates and Optimal Patches}
\usage{
r_env_CT(simul, n_patches, sp1, time)
}
\arguments{
\item{simul}{a simulation output from \code{multigen()}}

\item{n_patches}{the number of patches in the simulation}

\item{sp1}{the number of species}

\item{time}{the time slice at which e}
}
\value{
Outputs 4 data.frame of species by patch mismatches each with four columns:
\itemize{
\item \code{env_all} -> Environmental Growth only
\item \code{r_all}   -> Environmental + Hierarchical competition Growth only
\item \code{r_comp}  -> Abundance
\item \code{r_envab} -> Abundance based on environmental filtering only
}

The 4 data.frame are:
\itemize{
\item \code{env}       -> Patch
\item \code{sp}        -> Species
\item \code{r_env}     -> Observed statistic (environmental growth, total growth,
abundance, environmental filtering abundance)
\item \code{max_r_env} -> Patch of maximum observed statistic
}
}
\description{
Extract Species by Patch Growth Rates and Optimal Patches
}
