% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetacommunityDynamicsFctsOikos.R
\name{rdet}
\alias{rdet}
\title{Function definition for deterministic model run with global dispersal}
\usage{
rdet(R, d, a, patches, species, data)
}
\description{
Function definition for deterministic model run with global dispersal
}
