% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetacommunityDynamicsFctsOikos.R
\name{rstoch}
\alias{rstoch}
\title{Function definition for stochastic model run with global dispersal}
\usage{
rstoch(R, d, a, patches, species, data)
}
\description{
Function definition for stochastic model run with global dispersal
}
