% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{scale_distance}
\alias{scale_distance}
\title{Scale distance or matrix between 0 and 1}
\usage{
scale_distance(dist_matrix)
}
\arguments{
\item{dist_matrix}{a matrix or a dist object}
}
\description{
dist - min(dist) / (max(dist) - min(dist))
}
