% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_ab_growth_rates.R
\name{sp_ab_gr}
\alias{sp_ab_gr}
\title{Plot environmental response curves of species in all the patches}
\usage{
sp_ab_gr(simul, n_patches, sp, time)
}
\arguments{
\item{simul}{results array of a given simulation (a \code{\link[=multigen]{multigen()}} output)}

\item{n_patches}{an integer indicating the patch number}

\item{sp}{integer for the number of species to consider}

\item{time}{integer determining the time step of interest}
}
\description{
Automatically assigns one color per species.
}
