% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_gaussian.R
\name{truncated_gaussian}
\alias{truncated_gaussian}
\title{Truncated Gaussian Function}
\usage{
truncated_gaussian(t_opt, sigma_opt, a, b)
}
\arguments{
\item{t_opt}{[\code{numeric(1+)}]\cr{}
community optimal trait}

\item{sigma_opt}{[\code{numeric(1+)}]\cr{}
community optimal standard deviation}

\item{a}{[\code{numeric(1+)}]\cr{}
minimum value of trait in trait range}

\item{b}{[\code{numeric(1+)}]\cr{}
maximum value of trait in trait range}
}
\value{
the predicted CWM(s)
}
\description{
Returns the expected CWM of a community based on a truncated Gaussian as
proposed in Denelle et al. 2019 \url{doi:10.1111/oik.05851} (see Appendix 1,
OIK-05851). Under the hypothesis of a Gaussian environmental filter.
For each argument of the function the user can provide one or multiple values
as long as the vectors have the same lengths
}
