% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{wtd_kurtosis}
\alias{wtd_kurtosis}
\title{Weighted Kurtosis with na.rm}
\usage{
wtd_kurtosis(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{the values to be weighted}

\item{w}{the weighted}

\item{na.rm}{should NA be removed?}
}
\description{
Compute weighted kurtosis using \code{\link[Weighted.Desc.Stat:w.kurtosis]{Weighted.Desc.Stat::w.kurtosis}} but add an
option to remove \code{NA} values
}
