% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{wtd_skewness}
\alias{wtd_skewness}
\title{Weighted Skewness with na.rm}
\usage{
wtd_skewness(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{the values to be weighted}

\item{w}{the weighted}

\item{na.rm}{should NA be removed?}
}
\description{
Compute weighted Skewness using \code{\link[Weighted.Desc.Stat:w.skewness]{Weighted.Desc.Stat::w.skewness}} but add an
option to remove \code{NA} values
}
