% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_moments.R
\name{wtd_var}
\alias{wtd_var}
\title{Weighted Variance}
\usage{
wtd_var(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{values whose weighted mean is to be computed}

\item{w}{weights of the same length as \code{x} to be used against x}

\item{na.rm}{a logical values indicating whether \code{NA} values in both \code{x} and
\code{w} should be stripped before computation}
}
\description{
Weighted Variance
}
