% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{alphaterm}
\alias{alphaterm}
\title{Function definition for deterministic model run with global dispersal
Compute alpha term in Beverton-Holt function}
\usage{
alphaterm(distance, Nts, A, B)
}
\arguments{
\item{distance}{competiton matrix of species}

\item{Nts}{vector of abundances of species at time t}

\item{A}{scalar for the inter-specific competition}

\item{B}{scalar for the intra-specific competition}
}
\description{
From a competition matrix (for the moment the distance between species
traits) and a vector of abundances by species, return the alpha term in the
Beverton-Holt equation. The order of species between the two should be the
same as no checks are done.
Typically the competition matrix is an euclidean trait distance matrix
between species. The closer the species are the higher the combination.
The term is computed as follow:
}
\details{
\deqn{
    \alpha_i = \sum_{j = 1, j \neq i}^{S} N_{t, j, x} \times (1 -
                                                              \delta_{ij})
}{
    alpha_i = sum_{j = 1, j != i}^{S} Ntjx * (1 - delta_ij)
},
where alpha_i is the competition term of species i; Ntjx the abundance of
species j, at time t, in patch x and delta_ij the functional distance between
species i and species j.
}
