% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{bevHoltFct}
\alias{bevHoltFct}
\title{Beverton-Holt function}
\usage{
bevHoltFct(R, N, alpha, K)
}
\arguments{
\item{R}{a numeric vector of species growth rates}

\item{N}{a numeric vector of species population sizes}

\item{alpha}{the competition coefficient see \code{\link[=alphaterm]{alphaterm()}} for its
computation}

\item{K}{a numeric vector of species carrying capacity.
NB: This is a fixed property for all species and not of the site}
}
\description{
To simulate growth easily, use the Beverton-Holt equation. Which is:
}
\details{
\deqn{
   N_{t+1, i, x} = \frac{R_{i, x} \times N_{t, i, x}}{1 + A \times \alpha}
}{
   N(t+1) = (R * N(t))/(1 + A * alpha)
}
where t is time, i is the species of interest and x the patch it occupies.
}
