% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{compute_hierarchical_compet}
\alias{compute_hierarchical_compet}
\title{Compute Hierarchical Competition coefficient at each time step}
\usage{
compute_hierarchical_compet(
  composition_given_time_step,
  trait_values,
  trait_weights,
  H,
  exponent = 1
)
}
\arguments{
\item{composition_given_time_step}{composition matrix at a given time step
(a site-species matrix with sites in rows)}

\item{trait_values}{a trait matrix}

\item{trait_weights}{a scenario data.frame}

\item{H}{the hierarchical competition scalar}

\item{exponent}{exponent to use for hierarchical compet. distances}
}
\description{
Outputs a matrix of additional growth per patch per species given by
hierarchical competition. The values are considered
}
