% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weighted_performance.R
\name{compute_weighted_performance}
\alias{compute_weighted_performance}
\title{Community Performance Estimators}
\usage{
compute_weighted_performance(perf_df, trait_df)
}
\arguments{
\item{perf_df}{species performance data frame output by
\code{\link[=extract_performances_from_simul]{extract_performances_from_simul()}}}

\item{trait_df}{trait data.frame containing columns \code{trait_cor}, \code{seed} and
\code{species}}
}
\description{
Compute for each community all traits:
\itemize{
\item trait of best monoculture
\item trait of best polyculture
\item trait CWM
\item weighted trait by monoculture growth
\item weighted trait by polyculture growth
}
}
