% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{env_curve}
\alias{env_curve}
\title{Species growth rate for a given trait and environment}
\usage{
env_curve(trti, envx, types, k = 2, width = 0.5)
}
\arguments{
\item{trti}{a numeric vector of species trait values}

\item{envx}{a}

\item{types}{a character vector of trait types indicating if traits should be
used to compute growth rate (types \code{RA} or \code{R})}

\item{k}{a scalar giving the maximum growth rate in optimal environment}

\item{width}{a numeric for niche breadth, constant in gaussian function}
}
\description{
Using traits that affect growth rate and specified environments, this
function returns a numeric value of expected growth rates given the traits
and the environmental value. The total growth rate is then the average of the
growth rates computed with each trait.
}
\details{
For the moment the environmental filter follows a Gaussian distribution:

\deqn{
    R_{i, x} = k \times \exp(- \frac{(trait_i - env_x)^2}{2\times {width}^2})
}{
    R_ix = k * exp((t_i - env_x)^2/(2 * width^2))
},
where t_i is trait of species i, env_x the environmental value in patch x, k
a scalar giving the maximal growth rate and width the environmental breadth
of species.
}
