% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{landscape_df}
\alias{landscape_df}
\title{Simpler data.frame extracted from population}
\usage{
landscape_df(given_composition, trait_df, trait_weights, time = 150)
}
\arguments{
\item{given_composition}{a result of a call to \code{multigen()}}

\item{trait_df}{the trait data.frame}

\item{trait_weights}{the data.frame of trait weights}
}
\value{
a data.frame with the following columns
* \code{N} for population
* \code{species}
* \code{growth_trait}
}
\description{
Simpler data.frame extracted from population
}
