% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{multigen}
\alias{multigen}
\title{Function to run the simulation}
\usage{
multigen(traits, trait_type, env, time, species, patches, composition, A, d, k,
  width)
}
\arguments{
\item{traits}{a species-traits data.frame with species as rownames and
traits as numeric columns with names matching \code{trait_type}}

\item{trait_type}{a character vector indicating trait types (\code{R}=contributing
to growth only, \code{A} = contributing to competition only,
\code{RA} = contributing to both growth and competition or
\code{N} = trait not contributing to growth nor competition)}

\item{env}{a vector of environmental values}

\item{time}{an integer giving the total number of generations}

\item{species}{an integer giving the total number of species to simulate}

\item{patches}{an integer giving the total number of patches to simulate}

\item{composition}{the actual array containing species abundances per site
over time, giving the initial populations of each species}

\item{A}{the scalar of competition coefficent (see \code{\link[=bevHoltFct]{bevHoltFct()}})}

\item{d}{a numeric value between 0 and 1 giving the percentage of
dispersal occurring across all patches}

\item{k}{a scalar giving the maximum growth rate in optimal
environment}

\item{width}{a numeric giving niche breadth of all species}
}
\description{
Using specified parameters this function run the simulation
}
