% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigen_function.R
\name{multigen}
\alias{multigen}
\title{Function to run the simulation}
\usage{
multigen(traits, trait_type, env, time, species, patches, composition, A, d, k,
  c)
}
\arguments{
\item{traits}{a species-traits data.frame}

\item{trait_type}{a character vector indicating trait types (\code{R}, \code{A}, \code{RA}
or \code{N})}

\item{env}{a vector of environmenal values}

\item{time}{a integer giving the number of generations}

\item{species}{?}

\item{patches}{the number of patches?}

\item{composition}{?}

\item{A}{the scalar of competition coefficent (see \code{\link[=bevHoltFct]{bevHoltFct()}})}

\item{d}{?}

\item{k}{a scalar for computation fo growth rate}

\item{c}{a constant to compute growth rates}
}
\description{
Using specified parameters this function run the simulation
}
