% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_moments.R
\name{wtd_skewness}
\alias{wtd_skewness}
\title{Weighted Skewness with na.rm}
\usage{
wtd_skewness(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{values whose weighted mean is to be computed}

\item{w}{weights of the same length as \code{x} to be used against x}

\item{na.rm}{a logical values indicating whether \code{NA} values in both \code{x} and
\code{w} should be stripped before computation}
}
\description{
Compute weighted Skewness using \code{\link[Weighted.Desc.Stat:w.skewness]{Weighted.Desc.Stat::w.skewness}} but add an
option to remove \code{NA} values
}
